rename sdsandbox folder
