+++ examples/models/models.py
@@ -11,10 +11,10 @@ from tensorflow.keras.layers import Dropout, Flatten, Dense
 from tensorflow.keras.optimizers import Adam
 
 try:
-    import sdsandbox.conf as conf
+    import sdsandbox_perturbations.conf as conf
 except:
     print("trying to resolve conf import through relative path")
-    from ..sdsandbox import conf
+    from ..sdsandbox_perturbations import conf
 
 def show_model_summary(model):
     model.summary()
