+++ examples/models/train_dave2.py
@@ -20,7 +20,7 @@ project_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(project_root)
 
 # Now you can use absolute imports
-from sdsandbox import conf
+from sdsandbox_perturbations import conf
 from models import get_nvidia_model
 
 """
