--- examples/sdsandbox/requirements.txt
@@ -1,5 +1,0 @@
-numpy
-pillow
-h5py
-gym
--e git+https://github.com/tawnkramer/gym-donkeycar.git#egg=gym_donkeycar
