+++ examples/sdsandbox_perturbations/requirements.txt
@@ -1,0 +1,5 @@
+numpy
+pillow
+h5py
+gym
+-e git+https://github.com/tawnkramer/gym-donkeycar.git#egg=gym_donkeycar
