+++ examples/sdsandbox_perturbations/setup.sh
@@ -1,0 +1,4 @@
+curl -O https://syncandshare.lrz.de/dl/fi4EZwyMzgYMXcrb2EsGeJ/mymodel.h5
+curl -O https://syncandshare.lrz.de/dl/fiRXQY7i5k3SMh9uQDmS1i/model_1_11.h5
+curl -O https://syncandshare.lrz.de/dl/fiL4ocNC2vwTL32m8J1JB9/model_15_10.h5
+curl -O https://syncandshare.lrz.de/getlink/fiD9MBdQbQogMCuk5LDM3J/generatedRoadModel.h5
